Final fixes to atoi and atof conversion functions, moved all free functions to relevant files, moved validation functions to relevant files, added updated suppression file
