+++ include/miniRT.h
@@ -6,7 +6,7 @@
 /*   By: henbuska <henbuska@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/13 14:10:16 by irychkov          #+#    #+#             */
-/*   Updated: 2025/03/04 21:54:01 by henbuska         ###   ########.fr       */
+/*   Updated: 2025/03/05 12:55:49 by henbuska         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -101,6 +101,7 @@ double			validate_dimension(char *str);
 double			ft_atof(char *str);
 double			ft_atof_error(char *str, int *error);
 int				rt_atoi(char *str);
+int				handle_sign(char **str);
 t_tuple			string_to_point(char **strings);
 t_tuple			string_to_vector(char **strings);
 t_tuple			string_to_color(char **strings);
@@ -109,6 +110,7 @@ int				free_and_return(t_rt *rt, char **array, int ret);
 void			free_arrays(char **array1, char **array2);
 void			print_error(char *message);
 int				error(char *message, int ret);
+int				counter(char **s);
 
 /******************** MATH FOLDER **********/
 // Tuple
