+++ scenes/too_many_elements.rt
@@ -1,6 +1,8 @@
 A 0.2                                   255,255,255
-C 0,0,10hi       0,0,0,   70
-L -20,0,30                 0.7          255,255,255
+C 0,0,10       0,0,-1,   70
+L 0,10,10                 0.7          255,255,255
+L 0,10,10                 0.7          255,255,255
+
 pl 0,0,0          0,1,0                 255,0,225
-sp 0,0,20                  20           255,0,0
-cy 50.0,0.0,20.6  0,0.1,0  14.2  21.42  10,0,255
+sp 0,0,2                  0.3           255,0,0
+cy 5.0,0.0,2.6  0,0.1,0  4.2  2.42     10,0,255
